tidy: Convert directly from string to float instead of string to double to float.
Use the QString::toFloat function instead of the QString::toDouble version.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/narrowing-conversions.html

+++ mythtv/libs/libmythtv/programdata.cpp
@@ -479,7 +479,7 @@ uint DBEvent::GetOverlappingPrograms(
             query.value(7).toUInt(),
             query.value(8).toUInt(),
             query.value(9).toUInt(),
-            query.value(19).toDouble(),
+            query.value(19).toFloat(),
             query.value(10).toString(),
             query.value(11).toString(),
             query.value(18).toUInt(),
